fix: address Copilot review round 4
- Switch stdin from line-buffered to raw (character-at-a-time) mode so
  the userspace shell receives each keystroke immediately. Enter sends
  '\n', backspace sends 0x7f, Ctrl-C sends 0x03 — all handled by the
  shell, not the kernel. Removes line buffer, flush_line, backspace.
- Add checked_add guards in ELF relocation parsing to prevent overflow
  from malformed PT_DYNAMIC or .rela.dyn entries.
- Fix nanosleep timespec type: use i64 array (matching kernel's expected
  tv_sec/tv_nsec layout) instead of u64.

+++ userspace/syscall-lib/src/lib.rs
@@ -324,7 +324,7 @@ pub fn nanosleep(seconds: u64) -> isize {
     // The kernel's nanosleep reads a timespec struct from a user pointer:
     //   bytes 0..8: tv_sec (i64)
     //   bytes 8..16: tv_nsec (i64)
-    let ts: [u64; 2] = [seconds, 0];
+    let ts: [i64; 2] = [seconds as i64, 0];
     unsafe { syscall2(SYS_NANOSLEEP, ts.as_ptr() as u64, 0) as isize }
 }
